<ДЗ №11> - добавил возможность выйти из аккаунта
