+++ src/main/java/com/geekbrains/ru/springmvcdemo/config/SecurityConfig.java
@@ -2,6 +2,7 @@ package com.geekbrains.ru.springmvcdemo.config;
 
 import com.geekbrains.ru.springmvcdemo.service.impl.UserServiceImpl;
 import lombok.Data;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
@@ -15,6 +16,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 @Configuration
 @EnableWebSecurity //отключает стандартные настройки безопасности Spring Security и начинает использовать правила, прописанные в SecurityConfig
 @EnableGlobalMethodSecurity(securedEnabled = true, prePostEnabled = true) //активирует возможность ставить защиту на уровне методов (для этого над методами ставятся аннотации @Secured и @PreAuthorized).
+@RequiredArgsConstructor
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
     private final UserServiceImpl userService;
@@ -23,18 +25,17 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     protected void configure(HttpSecurity http) throws Exception {
         http
                 .authorizeRequests()
-                //Доступ только для не зарегистрированных пользователей
-                .antMatchers("/registration").not().fullyAuthenticated()
-                .antMatchers("/","/product").permitAll()
+                .antMatchers("/auth_page/**").authenticated()
+                .antMatchers("/product").permitAll()
                     .antMatchers("/admin/**").hasRole("ADMIN")
-                //Все остальные страницы требуют аутентификации
-                .anyRequest().authenticated()
+                .anyRequest().permitAll()
                     .and()
-                .formLogin().loginPage("/login").permitAll()
+                .formLogin()
+                .loginPage("/login").permitAll()
                 .failureUrl("/login-error")
                     .and()
                 .logout().permitAll()
-                .logoutSuccessUrl("/list")
+                .logoutSuccessUrl("/product").permitAll()
                 ;
     }
     //Объекты создаются как синглтоны
